API client: Refactor environments into their own file
