+++ firehydrant/types.go
@@ -136,31 +136,6 @@ type ServicesResponse struct {
 	Services []ServiceResponse `json:"data"`
 }
 
-// EnvironmentResponse is the payload for a single environment
-// URL: GET https://api.firehydrant.io/v1/environments/{id}
-type EnvironmentResponse struct {
-	ID          string    `json:"id"`
-	Name        string    `json:"name"`
-	Description string    `json:"description"`
-	Slug        string    `json:"slug"`
-	CreatedAt   time.Time `json:"created_at"`
-	UpdatedAt   time.Time `json:"updated_at"`
-}
-
-// CreateEnvironmentRequest is the payload for creating a service
-// URL: POST https://api.firehydrant.io/v1/services
-type CreateEnvironmentRequest struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
-
-// UpdateEnvironmentRequest is the payload for updating a environment
-// URL: PATCH https://api.firehydrant.io/v1/environments/{id}
-type UpdateEnvironmentRequest struct {
-	Name        string `json:"name,omitempty"`
-	Description string `json:"description,omitempty"`
-}
-
 // FunctionalityResponse is the payload for a single environment
 // URL: GET https://api.firehydrant.io/v1/functionalities/{id}
 type FunctionalityResponse struct {
